@@ -1,7 +1,7 @@
 {
-  "name": "dsj-events",
+  "name": "dsj_events",
   "version": "1.0.0",
-  "description": "My impressive Nuxt.js project",
+  "description": "My superb Nuxt.js project",
   "author": "Chris Halley",
   "private": true,
   "scripts": {
@@ -12,29 +12,21 @@
     "start": "nuxt start",
     "generate": "nuxt generate",
     "lint": "eslint --ext .js,.vue --ignore-path .gitignore .",
-    "precommit": "npm run lint",
-    "test": "mocha tests/*.test.js",
-    "test-watch": "nodemon --exec 'npm test'"
+    "precommit": "npm run lint"
   },
   "dependencies": {
     "@nuxtjs/axios": "^5.0.0",
-    "cloudinary-core": "^2.5.0",
+    "cross-env": "^5.2.0",
     "js-cookie": "^2.2.0",
     "jsonwebtoken": "^8.4.0",
-    "nuxt": "^1.0.0",
+    "nuxt": "^2.0.0",
     "vuelidate": "^0.7.4"
   },
   "devDependencies": {
-    "@vue/test-utils": "^1.0.0-beta.25",
+    "nodemon": "^1.11.0",
     "babel-eslint": "^8.2.1",
-    "concurrently": "^3.5.1",
-    "cross-env": "^5.0.1",
-    "eslint": "^4.15.0",
+    "eslint": "^5.0.1",
     "eslint-loader": "^2.0.0",
-    "eslint-plugin-vue": "^4.0.0",
-    "jest": "^23.6.0",
-    "node-sass": "^4.9.0",
-    "nodemon": "^1.18.6",
-    "sass-loader": "^7.0.1"
+    "eslint-plugin-vue": "^4.0.0"
   }
 }
